requirements.txt
openai-whisper
numpy==1.26.4
langchain
rich
langchain-community
soundfile
litellm